:ambulance: :: (#12) Move packages and resolve join issues
