+++ ida-secondary-domain/src/main/java/kr/hs/dgsw/cns/aggregate/secondary/domain/value/Examinee.java
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.cns.aggregate.secondary.domain.domain.value;
+package kr.hs.dgsw.cns.aggregate.secondary.domain.value;
 
 import kr.hs.dgsw.cns.aggregate.admission.domain.admission.Admission;
 import lombok.AllArgsConstructor;
