+++ ida-secondary-infrastructure/src/main/java/kr/hs/dgsw/cns/aggregate/secondary/entity/value/ExamineeVO.java
@@ -1,7 +1,8 @@
-package kr.hs.dgsw.cns.aggregate.secondary.value;
+package kr.hs.dgsw.cns.aggregate.secondary.entity.value;
 
 import jakarta.persistence.Embeddable;
-import kr.hs.dgsw.cns.aggregate.admission.domain.admission.Admission;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
 import kr.hs.dgsw.cns.aggregate.admission.entity.admission.AdmissionEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -19,6 +20,8 @@ public class ExamineeVO implements Serializable {
     @Serial
     private static final long serialVersionUID = -2238420653727913619L;
 
+    @OneToOne
+    @JoinColumn(name = "admission_id")
     private AdmissionEntity admission;
 
     private Long examCode;
